backend/src/requirements.txt
fastapi==0.111.0
uvicorn==0.15.0
aiohttp==3.9.5
pydantic-settings==2.3.4

pytube==15.0.0
youtube_comment_downloader==0.1.76
youtube-transcript-api==0.6.2